poker_backend/pyproject.toml
[tool.poetry]
name = "poker_backend"
version = "0.1.0"
description = "Backend for poker game"
authors = ["Nathnael Dereje <nathandere1357@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.111.0"
pydantic = "2.7.4"
uvicorn = "0.30.1"
asyncpg = "^0.29.0"
pokerkit = "^0.5.4"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.5.2"

[tool.poetry.group.dev.dependencies]
httpx = "0.27.0"
mypy = "1.10.0"
ruff = "0.4.8"

[build-system]
requires = ["poetry-core"]

build-backend = "poetry.core.masonry.api"
